Fixed Image Detection with Mutliple Samples
