+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/test/AutonomousTest.java
@@ -1,0 +1,54 @@
+package org.firstinspires.ftc.teamcode.opmodes.test;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+import org.firstinspires.ftc.teamcode.hardware.Robot;
+import org.firstinspires.ftc.teamcode.input.ControllerMap;
+import org.firstinspires.ftc.teamcode.opmodes.LoggingOpMode;
+import org.firstinspires.ftc.teamcode.opmodes.auto.AutonomousTemplate;
+import org.firstinspires.ftc.teamcode.util.Status;
+import org.firstinspires.ftc.teamcode.util.event.EventBus;
+
+// we going to use the event bus system for this so that everything can be done on one thread
+@Autonomous(name="Autonomous Test")
+public class AutonomousTest extends LoggingOpMode
+{
+    private Robot robot;
+    private AutonomousTemplate auto;
+    private String name = "Autonomous Test";
+    private int id;
+
+    @Override
+    public void init() {
+        super.init();
+        this.robot = Robot.initialize(hardwareMap, name);
+        this.auto = new AutonomousTemplate(
+                name,
+                this.robot,
+                hardwareMap,
+                new ControllerMap(gamepad1, gamepad2, new EventBus()),
+                telemetry
+        );
+        auto.init_camera();
+//        auto.init_server();
+//        auto.init_odometry(0, 0, 0);
+    }
+
+    @Override
+    public void start() {
+        robot.odometry.resetEncoders();
+    }
+
+    @Override
+    public void loop() {
+        auto.check_image();
+
+        telemetry.addData("Shipping Height: ", auto.shipping_height);
+        id = auto.update();
+    }
+
+    @Override
+    public void stop() {
+        auto.stop();
+    }
+}
